% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildExtendedEpiDataset.R
\name{BuildExtendedEpiDataset}
\alias{BuildExtendedEpiDataset}
\title{BuildExtendedEpiDataset
This function allows you to build an extended dataset of the standard epidataset (with log values, spline values, 3 days average,...)}
\usage{
BuildExtendedEpiDataset(country)
}
\arguments{
\item{Select}{the country to build the dataset for}
}
\description{
BuildExtendedEpiDataset
This function allows you to build an extended dataset of the standard epidataset (with log values, spline values, 3 days average,...)
}
